+++ src/media.rs
@@ -1,5 +1,6 @@
 pub(crate) mod audio_data;
 pub(crate) mod custom_audio_source;
+pub(crate) mod custom_audio_track;
 pub(crate) mod native_vad;
 pub(crate) mod video_frame;
 pub(crate) mod virtual_camera_device;
@@ -8,6 +9,7 @@ pub(crate) mod virtual_speaker_device;
 
 pub(crate) use audio_data::PyAudioData;
 pub(crate) use custom_audio_source::PyCustomAudioSource;
+pub(crate) use custom_audio_track::PyCustomAudioTrack;
 pub(crate) use native_vad::PyNativeVad;
 pub(crate) use video_frame::PyVideoFrame;
 pub(crate) use virtual_camera_device::PyVirtualCameraDevice;